requirements.txt
streamlit>=1.28.0
ortools>=9.7.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
anthropic>=0.18.0
polyline>=2.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
